btn moved to the correct component and syncs correctly
