+++ ui/component/commentCreate/view.jsx
@@ -28,9 +28,7 @@ import type { ElementRef } from 'react';
 import UriIndicator from 'component/uriIndicator';
 import usePersistedState from 'effects/use-persisted-state';
 import WalletTipAmountSelector from 'component/walletTipAmountSelector';
-import { COMMENT_SERVER_API, COMMENT_SERVER_NAME } from 'config';
 import { getStripeEnvironment } from 'util/stripe';
-import Comments from 'comments';
 const stripeEnvironment = getStripeEnvironment();
 
 const TAB_FIAT = 'TabFiat';
@@ -64,9 +62,6 @@ type Props = {
   sendTip: ({}, (any) => void, (any) => void) => void,
   setQuickReply: (any) => void,
   toast: (string) => void,
-  customCommentServers: Array<CommentServerDetails>,
-  setCommentServer: (string) => void,
-  commentServer: string,
 };
 
 export function CommentCreate(props: Props) {
@@ -91,13 +86,8 @@ export function CommentCreate(props: Props) {
     onDoneReplying,
     sendTip,
     setQuickReply,
-    customCommentServers,
-    setCommentServer,
-    commentServer,
   } = props;
 
-  const defaultServer = { name: COMMENT_SERVER_NAME, url: COMMENT_SERVER_API };
-  const allServers = [defaultServer, ...customCommentServers];
   const formFieldRef: ElementRef<any> = React.useRef();
   const buttonRef: ElementRef<any> = React.useRef();
 
@@ -504,29 +494,6 @@ export function CommentCreate(props: Props) {
             type={advancedEditor ? 'markdown' : 'textarea'}
             textAreaMaxLength={FF_MAX_CHARS_IN_COMMENT}
           />
-
-          <FormField
-            label="server"
-            type="select-tiny"
-            onChange={function (x) {
-              const selectedServer = x.target.value;
-              setCommentServer(selectedServer);
-              if (selectedServer === defaultServer.url) {
-                Comments.setServerUrl(undefined);
-              } else {
-                Comments.setServerUrl(selectedServer);
-              }
-            }}
-            value={commentServer}
-          >
-            {allServers.map(function (server) {
-              return (
-                <option key={server.url} value={server.url}>
-                  {server.name}
-                </option>
-              );
-            })}
-          </FormField>
         </>
       )}
